+++ CHANGELOG.md
@@ -30,7 +30,7 @@ All notable changes to this project will be documented in this file.<br>
 
 ## Master
 ### Added
-- Added `init?(iso8601:)` and `iso8601()` functions to create a Date from an ISO 8601 String and vice-versa in Date extension
+- Added `init?(iso8601:)` and `iso8601()` functions to create a Date from an ISO 8601 String and vice-versa in Date & String extensions
 - Added new functions in NSAttributedString extension to handle left, right, center and justified text alignment
 - Added `smoothLinearGradient(colors:, direction:)` function in UIView extension that uses `CGGradient` instead of `CAGradientLayer`
 - Added iPad Pro 12.9 2nd Generation and iPad Pro 10.5 devices in UIDevice extension